requirements.txt
gTTS>=2.3.1
moviepy>=1.0.3
numpy>=1.21.0
Pillow>=9.0.0
pydub>=0.25.1
requests>=2.28.0
SpeechRecognition>=3.8.1
python-dotenv>=0.19.0